Split status markdown fixtures
